Refactored tests for modularity
